updates checkout page with total of order, refactors location of calculateOrderAmount
